refactor: move all AuthenticationServiceExceptions to a new kt file
